+++ src/Amadeus/Client/RequestOptions/DocIssuanceIssueTicketOptions.php
@@ -22,6 +22,8 @@
 
 namespace Amadeus\Client\RequestOptions;
 
+use Amadeus\Client\RequestOptions\DocIssuance\CompoundOption;
+
 /**
  * DocIssuanceIssueTicketOptions
  *
@@ -31,26 +33,16 @@ namespace Amadeus\Client\RequestOptions;
 class DocIssuanceIssueTicketOptions extends Base
 {
     const PAX_TYPE_ADULT = "A";
-
     const PAX_TYPE_INFANT = "IN";
 
-
     const OPTION_ETICKET = "ET";
-
     const OPTION_RETRIEVE_PNR = "RT";
-
     const OPTION_ITINERARY_RECEIPT = "ITR";
-
     const OPTION_PRE_ISSUE_VALIDATION = "TKT";
-
     const OPTION_PRINT_ITIN = "IBP";
-
     const OPTION_PRINT_JOINT_ITIN = "IBJ";
-
     const OPTION_PRINT_EXTENDED_ITIN = "IEP";
-
     const OPTION_PRINT_EXTENDED_JOINT_ITIN = "IEJ";
-
     const OPTION_TICKET_ONLY = "TKA";
 
     /**
@@ -166,4 +158,11 @@ class DocIssuanceIssueTicketOptions extends Base
      * @var string
      */
     public $passengerType;
+
+    /**
+     * Compound Options
+     *
+     * @var CompoundOption[]
+     */
+    public $compoundOptions = [];
 }